research_assistant/pyproject.toml
[project]
name = "research_assistant"
version = "0.1.0"
description = "Multi-agent research assistant using context engineering principles"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[litellm,tools]==1.14.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "streamlit>=1.40.0",
    "firecrawl-py>=1.4.0",
    "zep-cloud>=2.0.0",
    # "zep-crewai>=0.1.0",  # incompatible with crewai 1.14+; using native zep-cloud v3 API instead
    "chromadb>=0.5.0",
    "pypdf>=4.0.0",
    # "tensorlake>=0.1.0",  # optional: install separately if PDF parsing is needed
    "sentence-transformers>=3.0.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
    "requests>=2.32.0",
]

[project.scripts]
research_assistant = "research_assistant.main:run"
run_crew = "research_assistant.main:run"
kickoff = "research_assistant.main:kickoff"
start_api = "research_assistant.main:start_api"
run_with_trigger = "research_assistant.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
